Hoist GraphicMap out of Map.py
